# api/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "1.0",
      "title": "Books Web API",
      "description": "Student Assignment API"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
...
    "paths": {
      "/books": {
        "get": {
          "tags": [
            "Product"
          ],
          "description": "Get all products in database, consisting of title, cost, price, average review score and image data.",
          "summary": "Get all products",
          "operationId": "getProducts",
          "parameters": [
            {
              "in": "query",
              "name": "title",
              "schema": {
                "type": "string"
              },
              "description": "Search by book's title"
            },
            {
              "$ref": "#/components/parameters/jsonParam"
            }
          ],
          "responses": {
            "200": {
              "description": "Get products successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
...
      "/publishers": {
        "get": {
          "tags": [
            "Publisher"
          ],
          "description": "Get list of all publishers from database.",
          "summary": "Get all publishers",
          "operationId": "getPublishers",
          "parameters": [
            {
              "$ref": "#/components/parameters/jsonParam"
            }
          ],
          "responses": {
            "200": {
              "description": "Get publishers successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Publisher"
                  }
                }
              }
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/publishers/{publisherId}": {
        "get": {
          "tags": [
            "Publisher"
          ],
          "description": "Get info of an existing publisher by Id.",
          "summary": "Get publisher's info by Id",
          "operationId": "getPublisherById",
          "parameters": [
            {
              "in": "path",
              "name": "publisherId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "Auto-generated ID of a publisher."
            },
            {
              "$ref": "#/components/parameters/jsonParam"
            }
          ],
          "responses": {
            "200": {
              "description": "Get publisher successfully."
            },
            "400": {
              "$ref": "#/components/responses/400"
            },
            "404": {
              "$ref": "#/components/responses/404"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/publishers/new": {
        "post": {
          "tags": [
            "Publisher"
          ],
          "description": "Add new publisher into database.",
          "summary": "Add new publisher",
          "operationId": "postPublisher",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/newPublisher"
                }
              }
            }
          },
          "responses": {
            "201": {
              "$ref": "#/components/responses/201"
            }
          }
        }
      },
...
      "/register": {
        "post": {
          "tags": [
            "User"
          ],
          "description": "Create a new account.",
          "summary": "Create account",
          "operationId": "postRegister",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/newRegister"
                }
              }
            }
          },
          "responses": {
            "201": {
              "$ref": "#/components/responses/201"
            },
            "405": {
              "$ref": "#/components/responses/405"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/login": {
        "post": {
          "tags": [
            "User"
          ],
          "description": "Log in an account.",
          "summary": "Login",
          "operationId": "postLogin",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string",
                      "format": "password"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "$ref": "#/components/responses/200"
            },
            "405": {
              "$ref": "#/components/responses/405"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
...
      "/languages": {
        "get": {
          "tags": [
            "Language"
          ],
          "description": "Get list of all supported language from database.",
          "summary": "Get languages",
          "operationId": "getLanguages",
          "parameters": [
            {
              "$ref": "#/components/parameters/jsonParam"
            },
            {
              "in": "query",
              "name": "name",
              "schema": {
                "type": "string"
              },
              "required": false,
              "description": "Optional. Get a list of languages containing the name."
            }
          ],
          "responses": {
            "200": {
              "$ref": "#/components/responses/200"
            },
            "405": {
              "$ref": "#/components/responses/405"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/languages/{languageId}": {
        "get": {
          "tags": [
            "Language"
          ],
          "description": "Get info of a language.",
          "summary": "View info of a language",
          "operationId": "getLanguage",
          "parameters": [
            {
              "in": "path",
              "name": "languageId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "Auto-generated ID of a language."
            }
          ],
          "responses": {
            "200": {
              "$ref": "#/components/responses/200"
            },
            "404": {
              "$ref": "#/components/responses/404"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/languages/new": {
        "post": {
          "tags": [
            "Language"
          ],
          "description": "Add new language to database.",
          "summary": "Add new language",
          "operationId": "postLanguage",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "$ref": "#/components/responses/201"
            },
            "401": {
              "$ref": "#/components/responses/Unauthorized"
            },
            "405": {
              "$ref": "#/components/responses/405"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "newPublisher": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "contact": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "string",
                  "format": "uri"
                },
                "phoneNumber": {
                  "type": "string",
                  "format": "uri"
                },
                "twitter": {
                  "type": "string",
                  "format": "uri"
                },
                "facebook": {
                  "type": "string",
                  "format": "uri"
                },
                "email": {
                  "type": "string",
                  "format": "email"
                }
              }
            },
            "origin": {
              "type": "string",
              "description": "Auto-generated ID of country"
            }
          }
        },
        "newRegister": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string"
            },
            "password": {
              "type": "string",
              "format": "password"
            },
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "gender": {
              "type": "string"
            },
            "address": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  }
                }
              }
            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "country": {
              "type": "string",
              "description": "Auto-generated ID of country."
            },
            "phoneNumber": {
              "type": "string"
            }
          },
          "required": [
            "username",
            "password",
            "firstName",
            "lastName",
            "gender",
            "email",
            "country",
            "phoneNumber"
          ]
        },
        "Author": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "middleName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "origin": {
              "type": "string"
            },
            "gender": {
              "type": "string"
            },
            "dateOfBirth": {
              "type": "string",
              "format": "date-time"
            },
            "biography": {
              "type": "string"
            }
          }
        },
        "Book": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "pageCount": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "author": {
              "$ref": "#/components/schemas/Author"
            },
            "language": {
              "$ref": "#/components/schemas/Language"
            },
            "genre": {
              "$ref": "#/components/schemas/BookGenre"
            },
            "coverType": {
              "$ref": "#/components/schemas/BookCover"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "image": {
              "type": "string",
              "format": "binary"
            },
            "imageType": {
              "type": "string"
            },
            "publishDate": {
              "type": "string",
              "format": "date-time"
            },
            "publisher": {
              "$ref": "#/components/schemas/Publisher"
            },
            "isbn": {
              "type": "object",
              "properties": {
                "isbn10": {
                  "type": "string"
                },
                "isbn13": {
                  "type": "string"
                }
              }
            }
          }
        },
        "BookCover": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string"
            }
          },
          "required": [
            "type"
          ]
        },
        "BookGenre": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            }
          },
          "required": [
            "name"
          ]
        },
        "Publisher": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "contact": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "string",
                  "format": "uri"
                },
                "phoneNumber": {
                  "type": "string",
                  "format": "uri"
                },
                "twitter": {
                  "type": "string",
                  "format": "uri"
                },
                "facebook": {
                  "type": "string",
                  "format": "uri"
                },
                "email": {
                  "type": "string",
                  "format": "email"
                }
              }
            },
            "origin": {
              "$ref": "#/components/schemas/Country"
            }
          },
          "required": [
            "name",
            "contact",
            "origin"
          ]
        },
        "Country": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "code": {
              "type": "string"
            }
          }
        },
        "Language": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "code": {
              "type": "string"
            }
          }
        },
        "User": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string"
            },
            "password": {
              "type": "string",
              "format": "password"
            },
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "gender": {
              "type": "string"
            },
            "address": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  }
                }
              }
            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "role": {
              "type": "string"
            },
            "country": {
              "$ref": "#/components/schemas/Country"
            },
            "phoneNumber": {
              "type": "string"
            }
          }
        },
        "Review": {
          "type": "object",
          "properties": {
            "product": {
              "$ref": "#/components/schemas/Product"
            },
            "review": {
              "type": "string"
            },
            "reviewer": {
              "$ref": "#/components/schemas/User"
            },
            "ratedScore": {
              "type": "integer",
              "minimum": 0
            }
          }
        },
        "Product": {
          "type": "object",
          "properties": {
            "detail": {
              "$ref": "#/components/schemas/Book"
            },
            "cost": {
              "type": "number",
              "format": "double"
            },
            "price": {
              "type": "number",
              "format": "double"
            },
            "review": {
              "$ref": "#/components/schemas/Review"
            }
          }
        },
...
      },
      "parameters": {
        "jsonParam": {
          "in": "query",
          "name": "json",
          "schema": {
            "type": "boolean"
          },
          "required": true,
          "description": "If true, return list in json format, otherwise render html page."
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "responses": {
        "200": {
          "description": "Request succeeded."
        },
        "201": {
          "description": "Resource created successfully."
        },
        "400": {
          "description": "Bad input"
        },
        "404": {
          "description": "Specific resource not found."
        },
        "405": {
          "description": "Method not allowed."
        },
        "500": {
          "description": "Internal server error"
        },
        "Unauthorized": {
          "description": "Unauthorized! Token not exist or invalid."
        }
      }
    }
  }